+++ src/libromdata/RomDataFactory.cpp
@@ -59,6 +59,7 @@ using std::vector;
 #include "WiiU.hpp"
 #include "EXE.hpp"
 #include "Nintendo3DS.hpp"
+#include "Nintendo3DSFirm.hpp"
 
 // Special case for Dreamcast save files.
 #include "dc_structs.h"
@@ -130,6 +131,7 @@ const RomDataFactoryPrivate::RomDataFns RomDataFactoryPrivate::romDataFns_header
 	GetRomDataFns(NES, false),
 	GetRomDataFns(WiiU, true),
 	GetRomDataFns(Nintendo3DS, true),
+	GetRomDataFns(Nintendo3DSFirm, false),
 
 	// NOTE: EXE has a 16-bit magic number,
 	// so it should go at the end.