+++ src/librpbase/RomData.cpp
@@ -458,6 +458,7 @@ const rp_char *RomData::fileType_string(void) const
 		_RP("Homebrew Application"),	// FTYPE_HOMEBREW
 		_RP("eMMC Dump"),		// FTYPE_EMMC_DUMP
 		_RP("Title Contents"),		// FTYPE_TITLE_CONTENTS
+		_RP("Firmware Binary"),		// FTYPE_FIRMWARE_BINARY
 	};
 	static_assert(ARRAY_SIZE(fileType_names) == FTYPE_LAST,
 		"fileType_names[] needs to be updated.");